+++ ui/core/src/index.ts
@@ -1,6 +1,6 @@
 export * from "./entities";
-export * from "./api";
-export * from "./actions";
+export * from "./services";
+export * from "./usecases";
 export * from "./store";
 export * from "./hooks";
 export * from "./config";
