+++ ui/core/src/services/EthbridgeService/PegTxEventEmitter.ts
