+++ ui/core/src/services/EthbridgeService/tokenContract.ts
