+++ ui/core/src/services/EthereumService/EthereumService.test.ts
