+++ ui/core/src/services/EthereumService/utils/getMetamaskProvider.ts
