+++ ui/core/src/test/utils/getMockWalletService.ts
@@ -1,4 +1,4 @@
-import { IWalletService } from "../../api/IWalletService";
+import { IWalletService } from "../../services/IWalletService";
 import { IAssetAmount } from "../../entities";
 
 export function getMockWalletService(
